requirements.txt
asgiref==3.5.0
Django==4.0.4
django-jazzmin==2.5.0
django-qr-code==3.0.0
djangorestframework==3.13.1
pytz==2022.1
segno==1.4.1
sqlparse==0.4.2
